Move loading the config for all projects to the /[project]/[version]/docs route
