+++ app/projects/query.tsx
@@ -1,5 +1,5 @@
 import { useRouteLoaderData } from '@remix-run/react'
-import type { QueryConfigLoader } from '~/routes/query.$version'
+import type { QueryConfigLoader } from '~/routes/query.$version.docs'
 
 export const gradientText =
   'inline-block text-transparent bg-clip-text bg-gradient-to-r from-red-500 to-amber-500'
@@ -41,7 +41,7 @@ export type Framework = 'angular' | 'react' | 'svelte' | 'vue' | 'solid'
 
 export const useQueryDocsConfig = () => {
   const queryConfigLoaderData = useRouteLoaderData<QueryConfigLoader>(
-    'routes/query.$version'
+    'routes/query.$version.docs'
   )
 
   if (!queryConfigLoaderData?.tanstackDocsConfig) {
