+++ app/projects/ranger.tsx
@@ -1,5 +1,5 @@
 import { useRouteLoaderData } from '@remix-run/react'
-import type { RangerConfigV1Loader } from '~/routes/ranger.v1'
+import type { RangerConfigV1Loader } from '~/routes/ranger.v1.docs'
 
 export const repo = 'tanstack/ranger'
 
@@ -9,11 +9,13 @@ export const gradientText =
   'inline-block text-transparent bg-clip-text bg-gradient-to-r from-lime-500 to-emerald-500'
 
 export const useRangerV1Config = () => {
-  const tableConfigLoaderData =
-    useRouteLoaderData<RangerConfigV1Loader>('routes/ranger.v1')
-  if (!tableConfigLoaderData?.tanstackDocsConfig) {
-    throw new Error('Config could not be read for tanstack/table!')
+  const rangerConfigLoaderData = useRouteLoaderData<RangerConfigV1Loader>(
+    'routes/ranger.v1.docs'
+  )
+
+  if (!rangerConfigLoaderData?.tanstackDocsConfig) {
+    throw new Error(`Config could not be read for ${repo}!`)
   }
 
-  return tableConfigLoaderData.tanstackDocsConfig
+  return rangerConfigLoaderData.tanstackDocsConfig
 }
